Moving sequence player utils.
