+++ axelrod/strategies/sequence_player.py
@@ -1,8 +1,5 @@
-import collections
-import functools
-import itertools
-
 from axelrod import Actions, Player, init_args
+from axelrod._strategy_utils import thue_morse_generator
 
 
 class SequencePlayer(Player):
@@ -35,57 +32,6 @@ class SequencePlayer(Player):
         # Be sure to reset the sequence generator
         Player.reset(self)
         self.sequence_generator = self.generator_function(*self.generator_args)
-
-
-class Memoized(object):
-   """Decorator. Caches a function's return value each time it is called.
-   If called later with the same arguments, the cached value is returned
-   (not reevaluated). From:
-   https://wiki.python.org/moin/PythonDecoratorLibrary#Memoize
-   """
-   def __init__(self, func):
-        self.func = func
-        self.cache = {}
-
-   def __call__(self, *args):
-        if not isinstance(args, collections.Hashable):
-            # uncacheable. a list, for instance.
-            # better to not cache than blow up.
-            return self.func(*args)
-        if args in self.cache:
-            return self.cache[args]
-        else:
-            value = self.func(*args)
-            self.cache[args] = value
-            return value
-
-   def __repr__(self):
-        """Return the function's docstring."""
-        return self.func.__doc__
-
-   def __get__(self, obj, objtype):
-        """Support instance methods."""
-        return functools.partial(self.__call__, obj)
-
-
-@Memoized
-def recursive_thue_morse(n):
-    """The recursive definition of the Thue-Morse sequence. The first few terms
-    of the Thue-Morse sequence are:
-    0 1 1 0 1 0 0 1 1 0 0 1 0 1 1 0 . . ."""
-
-    if n == 0:
-        return 0
-    if n % 2 == 0:
-        return recursive_thue_morse(n / 2)
-    if n % 2 == 1:
-        return 1 - recursive_thue_morse((n - 1) / 2)
-
-def thue_morse_generator(start=0):
-    """A generator for the Thue-Morse sequence."""
-
-    for n in itertools.count(start):
-        yield recursive_thue_morse(n)
 
 
 class ThueMorse(SequencePlayer):
